Introduced new API for working with metadata and enums
